Add parse test assertions
Add `assertParseTestFalse()` and `assertParseTestTrue()` to `ParseTest`.

Also, rename `evalTest()` to `assertParseTestResult()`. In the process, make
it a full-fledged assertion method.

The `assertParseTestResult()` assertion is the most generic. It requires a
`Test` to use as well as a result. The `assertParseTestTrue()` and
`assertParseTestFalse()` are very specific, using the `Test` object that comes
from `buildTest()` and asserting the result as true or false, respectively.