Relax rules for TestExitOrDie
Exit and die should both be safe if passed an integer or if no value is
passed in.

Integers are used for returning exit status to the calling process. Also, it
hasn't been sent to output since PHP 4.2.

Using a bare `exit` with no parameters is a valid way to stop the execution of
the script. Maybe not the best architecturally, but still valid.