Update composer dependencies to latest version. Fix #92
The previous composer.json was forcing the installation of an old
version
of the php parser, incompatible with the newer PHP 7.1 syntax. Now a PHP
7.1 project will be correctly parsed without errors.

The Symfony dependencies were also updated to allow a "composer require"
from a modern project using Symfony 4.

With these changes I had to fix part of the code for making the tests go
green again:

* implement all the methods of PhpParser\Node in FakeNode with correct
return type
* return correctly the name if node is a const
* drop the mocking of ProgressBar (from Symfony) because now the class
is final

I also fixed a typo in the comments :)